Move PkiUtils to appropriate package
