+++ src/jvmMain/kotlin/ehn/techiop/hcert/kotlin/crypto/CryptoAdapter.kt
@@ -1,7 +1,7 @@
 package ehn.techiop.hcert.kotlin.crypto
 
 
-import ehn.techiop.hcert.kotlin.chain.common.PkiUtils
+import ehn.techiop.hcert.kotlin.crypto.PkiUtils
 import ehn.techiop.hcert.kotlin.trust.ContentType
 import kotlinx.datetime.Clock
 import org.bouncycastle.asn1.pkcs.PrivateKeyInfo
